+++ src/periodicTable/FuelPrepopulator.java
@@ -1,4 +1,4 @@
-package fuels;
+package periodicTable;
 
 import java.util.HashMap;
